+++ src/lib/mem.c
@@ -7,10 +7,10 @@ char end_mem;
 #define SMEM (&start_mem)
 #define EMEM (&end_mem)
 
-static char* cur = SMEM;
+static char *cur = SMEM;
 
-void* simple_malloc(size_t size){
-    char* tmp;
+void *simple_malloc(uint32 size){
+    char *tmp;
 
     if((uint64)cur + size > (uint64)EMEM){
         uart_printf("[!] No enough space!\r\n");
